Add fixes for HW

+++ app/controllers/questions_controller.rb
@@ -1,9 +1,9 @@
 class QuestionsController < ApplicationController
   before_action :authenticate_user!, except: %i[index show]
-  before_action :get_new_answer, only: %i[show]
 
   expose :questions, -> { Question.all }
   expose :question
+  expose :answer, -> { question.answers.new }
 
   def update
     if question.update(question_params)
@@ -14,17 +14,21 @@ class QuestionsController < ApplicationController
   end
 
   def create
-    @exposed_question = current_user.created_questions.new(question_params)
-    if @exposed_question.save
-      redirect_to @exposed_question, notice: 'Your question successfully created'
+    @exposed_question = current_user.questions.new(question_params)
+    if question.save
+      redirect_to question, notice: 'Your question successfully created'
     else
       render :new
     end
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path, notice: 'Question deleted successfully.'
+    if current_user.author_of?(question)
+      question.destroy
+      redirect_to questions_path, notice: 'Question deleted successfully.'
+    else
+      render :show, alert: 'You can not delete not your own question'
+    end
   end
 
   private
@@ -33,8 +37,4 @@ class QuestionsController < ApplicationController
     params.require(:question).permit(:title, :body)
   end
 
-  def get_new_answer
-    @answer = session[:error_obj] ? session[:error_obj] : question.answers.new
-    session[:error_obj] = nil
-  end
 end
